requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
beautifulsoup4==4.12.2
newspaper3k==0.2.8
lxml[html_clean]==4.9.3
python-dateutil==2.8.2
python-dotenv==1.0.0
Pillow==10.1.0
django-extensions==3.2.3
tiktoken==0.5.2
tiktoken